+++ internal/apiserver/unblock_endpoint_test.go
@@ -26,8 +26,11 @@ type stubFirewallBackend struct {
 
 var _ firewall.Backend = (*stubFirewallBackend)(nil)
 
-func (s *stubFirewallBackend) EnsureBase() error                                 { return nil }
-func (s *stubFirewallBackend) AddBlock(net.IP, string, *time.Duration) error     { return nil }
+func (s *stubFirewallBackend) EnsureBase() error                             { return nil }
+func (s *stubFirewallBackend) AddBlock(net.IP, string, *time.Duration) error { return nil }
+func (s *stubFirewallBackend) AddBlockBatch([]firewall.BlockEntry) (firewall.BlockBatchResult, error) {
+	return firewall.BlockBatchResult{}, nil
+}
 func (s *stubFirewallBackend) ListBlocks() ([]firewall.BlockedEntry, error)      { return nil, nil }
 func (s *stubFirewallBackend) ListAllows() ([]firewall.BlockedEntry, error)      { return nil, nil }
 func (s *stubFirewallBackend) AddAllow(net.IP, *time.Duration) error             { return nil }